backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
requests==2.32.3
PyPDF2==3.0.1
# Optional for better arXiv parsing
feedparser==6.0.11
# Optional security libs (recommended for production)
# passlib[bcrypt]==1.7.4
# argon2-cffi==23.1.0
google-generativeai==0.7.2
pinecone-client==4.1.0
# Optional alternative vector DB
# chromadb==0.5.0
